refactor : debug 패키지에서 main 패키지로 이동
