+++ feature/mypage/src/main/kotlin/com/wepli/mypage/menus/devmode/network/main/component/TagComponent.kt
@@ -1,4 +1,4 @@
-package com.wepli.mypage.devmode.network.main.component
+package com.wepli.mypage.menus.devmode.network.main.component
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.border
