+++ feature/mypage/src/main/kotlin/com/wepli/mypage/menus/devmode/network/main/enums/ApiMethodUiTag.kt
@@ -1,4 +1,4 @@
-package com.wepli.mypage.devmode.network.main.enums
+package com.wepli.mypage.menus.devmode.network.main.enums
 
 import androidx.compose.ui.graphics.Color
 import debug.model.ApiMethod
