+++ feature/mypage/src/main/kotlin/com/wepli/mypage/menus/devmode/network/main/viewmodel/NetworkLogViewModel.kt
@@ -1,20 +1,19 @@
-package com.wepli.mypage.devmode.network.main.viewmodel
+package com.wepli.mypage.menus.devmode.network.main.viewmodel
 
-import androidx.lifecycle.viewModelScope
 import base.BaseMviViewModel
 import base.Intent
 import base.SideEffect
 import base.UiState
-import com.wepli.mypage.devmode.network.main.enums.ApiMethodUiTag
+import com.wepli.mypage.menus.devmode.network.main.enums.ApiMethodUiTag
 import dagger.hilt.android.lifecycle.HiltViewModel
 import debug.model.ApiLog
 import debug.repository.DebugApiLogRepository
 import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.launch
 import javax.inject.Inject
 
 
 data class NetworkLogState(
+    val maxApiLogs: Int = 50,
     val originApiLogs: List<ApiLog> = emptyList(),
     val filteredApiLogs: List<ApiLog> = emptyList(),
     val selectedTag: ApiMethodUiTag = ApiMethodUiTag.ALL,
@@ -34,10 +33,6 @@ class NetworkLogViewModel @Inject constructor(
     initialState = NetworkLogState()
 ) {
 
-    companion object {
-        private const val MAX_LOG = 50
-    }
-
     init {
         collectApiLog()
     }
@@ -50,7 +45,9 @@ class NetworkLogViewModel @Inject constructor(
 
     private fun collectApiLog() = intent {
         launch(Dispatchers.IO) {
-            val sortedLogs = apiLogRepository.getLogs(MAX_LOG).sortedByDescending { it.startTime }
+            val sortedLogs = apiLogRepository
+                .getLogs(state.maxApiLogs)
+                .sortedByDescending { it.startTime }
             val filteredApiLogs = filterLogs(sortedLogs, state.selectedTag)
 
             // 이전과 같으면 생략
